requirements.txt
Django==5.2.4
openai
gunicorn
python-decouple
psycopg[binary]
dj-database-url
requests
django-allauth
django-ratelimit
dodopayments
Pillow
tiktoken
djangorestframework